airbnb削除, react/all適用
