+++ packages/frontend/.eslintrc.json
@@ -8,15 +8,14 @@
   "extends": [
     "next/core-web-vitals",
     "eslint:recommended",
+    "plugin:react/all",
     // import
     "plugin:import/recommended",
     "plugin:import/typescript",
-    "airbnb",
     // typescript-eslintの設定解説→ https://zenn.dev/cybozu_frontend/articles/ts-eslint-v6-guide
     "plugin:@typescript-eslint/strict-type-checked",
-    "plugin:react/recommended",
-    // "plugin:react-hooks/recommended",
-    // prettierと競合するeslintの設定を除外する
+    "plugin:promise/recommended",
+    // prettierと競合するeslintの設定を除外する 末尾に入れておくこと
     "prettier"
   ],
   "parser": "@typescript-eslint/parser",
@@ -188,6 +187,18 @@
     ],
     "@typescript-eslint/no-floating-promises": "off",
     // react
+    "react/self-closing-comp": "warn",
+    // <div>Hoge</div> みたいなのをNGにしない
+    "react/jsx-no-literals": "off",
+    // JSXの深さ制限
+    "react/jsx-max-depth": [
+      "error",
+      {
+        "max": 5
+      }
+    ],
+    // style, classnameなどのpropsを禁止する MUIと相性悪すぎるので一旦停止
+    "react/forbid-component-props": "off",
     "react/jsx-no-useless-fragment": "off",
     "react/require-default-props": "off",
     "react/prop-types": "off",
