+++ packages/frontend/package.json
@@ -43,7 +43,6 @@
     "@vitejs/plugin-react": "4.1.0",
     "@vitest/coverage-v8": "0.34.6",
     "eslint": "8.46.0",
-    "eslint-config-airbnb": "^19.0.4",
     "eslint-config-next": "13.4.13",
     "eslint-config-prettier": "^9.0.0",
     "eslint-import-resolver-typescript": "^3.6.0",
