+++ packages/frontend/src/app/layout.tsx
@@ -1,4 +1,4 @@
-export default function RootLayout({ children }: { children: React.ReactNode }) {
+export default function RootLayout({ children }: { readonly children: React.ReactNode }) {
   return (
     <html lang="ja">
       <body>{children}</body>
